feat(all):added pre-commit and fix some bugs

+++ src/geminiplayground/cli.py
@@ -1,4 +1,5 @@
 import os
+
 import typer
 from typing_extensions import Annotated
 
@@ -11,23 +12,25 @@ def check_api_key():
     """
     # attempt to load the api key from the .env file
     from dotenv import load_dotenv, find_dotenv
+
     load_dotenv(find_dotenv())
     # receive the api key from the command line
     api_key = os.environ.get("AISTUDIO_API_KEY", None)
     if not api_key:
         typer.echo(
             "Please set the AISTUDIO_API_KEY environment variable, or create a .env file with the api key obtained "
-            "from https://aistudio.google.com/app/apikey")
+            "from https://aistudio.google.com/app/apikey"
+        )
         raise typer.Abort()
 
 
 @cli.command()
 def ui(
-        host: str = "localhost",
-        port: int = 8081,
-        workers: int = os.cpu_count() * 2 + 1,
-        reload: Annotated[bool, typer.Option("--reload")] = True,
-        timeout: int = 12600,
+    host: str = "localhost",
+    port: int = 8081,
+    workers: int = int(os.cpu_count() * 2 + 1),  # type: ignore
+    reload: Annotated[bool, typer.Option("--reload")] = True,
+    timeout: int = 12600,
 ):
     """
     Launch the web app
@@ -47,11 +50,10 @@ def ui(
 
 @cli.command()
 def api(
-        host: str = "localhost",
-        port: int = 8081,
-        workers: int = os.cpu_count() * 2 + 1,
-        reload: Annotated[bool, typer.Option("--reload")] = True
-
+    host: str = "localhost",
+    port: int = 8081,
+    workers: int = int(os.cpu_count() * 2 + 1),  # type: ignore
+    reload: Annotated[bool, typer.Option("--reload")] = True,
 ):
     """
     Launch the API
